review renaming, more javadoc
