+++ platform-sdk/consensus-otter-tests/src/testFixtures/proto/container_communication.proto
@@ -189,11 +189,12 @@ message TransactionRequest {
   bytes payload = 1;
 }
 
-message KillImmediatelyRequest {
-  google.protobuf.Empty ignore = 1;
+message TransactionRequestAnswer {
+  // indicator if the platform accepted the transaction
+  bool result = 1;
 }
 
-message KillImmediatelyAnswer {
+message KillImmediatelyRequest {
   google.protobuf.Empty ignore = 1;
 }
 
@@ -203,8 +204,8 @@ service TestControl {
   rpc Start(StartRequest) returns (stream EventMessage);
 
   // RPC used by the test harness to submit a transaction to the running
-  // platform. Returns an empty message to acknowledge the submission.
-  rpc SubmitTransaction(TransactionRequest) returns (google.protobuf.Empty);
+  // platform. Returns an TransactionRequestAnswer weather the platform accepted the transaction or not.
+  rpc SubmitTransaction(TransactionRequest) returns (TransactionRequestAnswer);
 
   // RCP to signal a kill of the app
   rpc KillImmediately(KillImmediatelyRequest) returns (google.protobuf.Empty);
